+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 
 * Bump the minimum required Emacs version to 25.1.
 * Drop support for Java 7 and Clojure(Script) 1.7.
+* Add support for printing to the current buffer to `cider-eval-defun-to-point`.
 
 ## 0.17.0 (2018-05-07)